refactor SetProperties in master.py to be aside SetProperty
